data collection and processing scripts, removing files as I do a directory restructuring
