--- 2.web_db_scripts/example-mLab-credentials.yml
@@ -1,4 +1,0 @@
-db_name: pathcore-demo-server
-mLab_URI: ds151068.mlab.com:51068
-password: <your db password>
-user: <your db username>
