+++ data/download_data.sh
@@ -47,7 +47,7 @@ curl $models_url | tar --extract --gzip --strip-components=1 --directory=$ensemb
 pseudomonas_KEGG_file=$pa_data_dir"/pseudomonas_KEGG_definitions.txt"
 
 # Uncomment this command to get an up-to-date version:
-# python get_annotations.py $pseudomonas_KEGG_file KEGG
+# python ./get_annotations_from_Tribe.py $pseudomonas_KEGG_file KEGG
 
 ##############################################################################
 # TCGA Pan-cancer dataset
@@ -68,3 +68,20 @@ curl $tcga_url | tar --extract --gzip --strip-components=1 --directory=$tcga_dat
 # with only HiSeqV2; instead, we just remove the other 2 files after download.
 rm $tcga_data_dir"/PANCAN_clinicalMatrix"
 rm $tcga_data_dir"/PANCAN_mutation"
+
+# Both the PAO1 and TCGA NMF models we used in our analysis are provided in
+# this repository. Because Github limits file size to <50 Mb, the TCGA NMF
+# model is provided as a .tar.gz and decompressed here.
+tcga_NMF_model_dir=$tcga_data_dir"/NMF_model"
+tcga_NMF_model_compressed=$tcga_NMF_model_dir"/NMF_300_features.tar.gz"
+tar -xzvf $tcga_NMF_model_compressed -C $tcga_NMF_model_dir
+
+# The PID pathway definitions file was generated by running the following
+# command. For versioning, the original `c2.cp.v6.0.symbols.gmt` file we
+# used in our analysis is provided in $tcga_data_dir. (See README for more
+# information)
+
+PID_file=$tcga_data_dir"/PID_pathway_definitions.txt"
+
+# Uncomment this command to get the file from `c2.cp.v6.0.symbols.gmt`
+# python get_PID_pathway_definitions.py ./tcga_data/c2.cp.v6.0.symbols.gmt $PID_file
